+++ src/main/java/com/kdang/task/controller/TaskController.java
@@ -1,10 +1,13 @@
 package com.kdang.task.controller;
 
 import com.kdang.task.model.Task;
+import com.kdang.task.model.TaskSpecification;
 import com.kdang.task.service.TaskService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @RestController
 @RequestMapping("/v1")
@@ -30,16 +33,25 @@ public class TaskController {
 
     // UPDATE
     @RequestMapping(value = "/tasks/{taskID}", method = RequestMethod.PUT)
-    public Task updateTask(@PathVariable(value = "taskID") Integer taskID,
+    public Task updateTask(@PathVariable(value = "taskID") Long taskID,
                            @RequestBody Task taskDetails) {
         return taskService.updateTask(taskID, taskDetails);
     }
 
     // DELETE
     @RequestMapping(value = "/tasks/{taskID}", method = RequestMethod.DELETE)
-    public void deleteTask(@PathVariable(value = "taskID") Integer taskID) {
+    public void deleteTask(@PathVariable(value = "taskID") Long taskID) {
         taskService.deleteTask(taskID);
     }
 
-
+//    @RequestMapping(value = "/tasks", method = RequestMethod.GET)
+//    @ResponseBody
+//    public List<Task> search(@RequestParam(value = "search") String search) {
+//        TaskSpecification builder = new TaskSpecification();
+//        Pattern pattern = Pattern.compile("(\\w+?)(:|<|>)(\\w+?),");
+//        Matcher matcher = pattern.matcher(search + ",");
+//        while (matcher.find()) {
+//            builder.with(matcher.group(1), matcher.group(2), matcher.group(3));
+//        }
+//    }
 }
